backend/requirements.txt
# --- Web API ---
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# --- Core CV & Math ---
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
pandas==2.1.3
scikit-learn==1.3.2

# --- Face Recognition (ArcFace) ---
insightface==0.7.3
onnxruntime==1.16.1

# --- YOLOv5-Face Detection ---
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
PyYAML>=6.0
tqdm>=4.66.0

# --- Tracking (ByteTrack wrapper) ---
supervision>=0.17.0

# --- Google Sheets / Drive ---
gspread==5.12.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
